feat: implement stack selection logic and tests for type compatibility
